requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.12
feedparser==6.0.11
httpx==0.28.1
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.32.3
uvicorn[standard]==0.34.1
yfinance==0.2.55
pytest==8.3.5
